+++ module-common/src/main/java/kitchenpos/menu/service/MenuGroupService.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.service;
 
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.menu.dto.MenuGroupRequest;
 import kitchenpos.menu.dto.MenuGroupResponse;
 import org.springframework.stereotype.Service;
