+++ module-common/src/main/java/kitchenpos/menu/service/MenuService.java
@@ -2,7 +2,7 @@ package kitchenpos.menu.service;
 
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.repository.MenuRepository;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
 import org.springframework.stereotype.Service;
