+++ module-common/src/main/java/kitchenpos/order/service/OrderStatusService.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.service;
 
 import kitchenpos.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.repository.OrderRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
