+++ module-repository-jpa/src/main/java/kitchenpos/repository/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.repository;
 
 import kitchenpos.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
