+++ module-repository-jpa/src/main/java/kitchenpos/repository/OrderLineItemRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.repository;
 
 import kitchenpos.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
