+++ module-repository-jpa/src/main/java/kitchenpos/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.repository;
 
 import kitchenpos.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
